@@ -1,90 +1,69 @@
 {
-  "name": "@coreui/coreui-free-react-admin-template",
-  "version": "3.0.0",
-  "description": "CoreUI React Open Source Bootstrap 4 Admin Template",
-  "author": {
-    "name": "CoreUI",
-    "url": "https://coreui.io",
-    "github": "https://github.com/coreui",
-    "twitter": "https://twitter.com/core_ui"
-  },
-  "contributors": [
-    {
-      "name": "CoreUI Team",
-      "url": "https://github.com/orgs/coreui/people"
-    }
-  ],
+  "name": "@kerti/balances",
+  "version": "4.4.0",
+  "description": "Balances UI - Frontend for Balances, a balance-based personal finance management tool",
   "homepage": ".",
-  "copyright": "Copyright 2017-2020 creativeLabs Łukasz Holeczek",
-  "license": "MIT",
-  "private": true,
+  "bugs": {
+    "url": "https://github.com/kerti/balances/issues"
+  },
   "repository": {
     "type": "git",
-    "url": "git@github.com:coreui/coreui-free-react-admin-template.git"
+    "url": "git@github.com:kerti/balances.git"
   },
-  "dependencies": {
-    "@coreui/chartjs": "^2.0.0",
-    "@coreui/coreui": "^3.2.2",
-    "@coreui/icons": "2.0.0-beta.4",
-    "@coreui/icons-react": "^1.0.2",
-    "@coreui/react": "^3.2.2",
-    "@coreui/react-chartjs": "^1.0.1",
-    "@coreui/utils": "^1.3.1",
-    "axios": "^0.21.1",
-    "classnames": "^2.2.6",
-    "cookies-js": "^1.2.3",
-    "core-js": "^3.6.5",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.3",
-    "humps": "^2.0.1",
-    "i18next": "^19.6.3",
-    "node-sass": "^4.14.1",
-    "normalizr": "^3.6.0",
-    "prop-types": "^15.7.2",
-    "react": "^16.13.1",
-    "react-app-polyfill": "^1.0.6",
-    "react-dom": "^16.13.1",
-    "react-i18next": "^11.7.0",
-    "react-redux": "7.2.0",
-    "react-router-dom": "^5.2.0",
-    "redux": "4.0.5",
-    "redux-devtools": "^3.6.0",
-    "redux-devtools-dock-monitor": "^1.1.3",
-    "redux-devtools-log-monitor": "^2.0.0",
-    "redux-logger": "^3.0.6",
-    "redux-thunk": "^2.3.0"
-  },
-  "devDependencies": {
-    "prettier": "2.0.5",
-    "react-scripts": "^3.4.1"
-  },
+  "license": "MIT",
+  "author": "Raditya Kertiyasa (https://github.com/kerti)",
   "scripts": {
-    "start": "react-scripts start",
     "build": "react-scripts build",
+    "changelog": "auto-changelog --starting-version 4.4.0 --commit-limit false --hide-credit",
+    "eject": "react-scripts eject",
+    "lint": "eslint \"src/**/*.js\"",
+    "start": "react-scripts start",
     "test": "react-scripts test",
     "test:cov": "npm test -- --coverage --watchAll=false",
-    "test:debug": "react-scripts --inspect-brk test --runInBand",
-    "eject": "react-scripts eject"
+    "test:debug": "react-scripts --inspect-brk test --runInBand"
   },
-  "bugs": {
-    "url": "https://github.com/coreui/coreui-free-react-admin-template/issues"
+  "dependencies": {
+    "@coreui/chartjs": "^3.0.0",
+    "@coreui/coreui": "^4.2.0",
+    "@coreui/icons": "^2.1.0",
+    "@coreui/icons-react": "^2.1.0",
+    "@coreui/react": "^4.3.0",
+    "@coreui/react-chartjs": "^2.1.0",
+    "@coreui/utils": "^1.3.1",
+    "@reduxjs/toolkit": "^1.8.3",
+    "axios": "^0.27.2",
+    "chart.js": "^3.8.0",
+    "classnames": "^2.3.1",
+    "cookies-js": "^1.2.3",
+    "core-js": "^3.23.4",
+    "humps": "^2.0.1",
+    "normalizr": "^3.6.2",
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
+    "react-app-polyfill": "^3.0.0",
+    "react-dom": "^18.2.0",
+    "react-i18next": "^11.18.3",
+    "react-redux": "^8.0.2",
+    "react-router-dom": "^6.3.0",
+    "redux": "^4.2.0",
+    "redux-logger": "^3.0.6",
+    "redux-thunk": "^2.4.1",
+    "simplebar-react": "^2.4.1"
   },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 10",
-    "not op_mini all"
-  ],
-  "jest": {
-    "collectCoverageFrom": [
-      "src/**/*.{js,jsx}",
-      "!**/*index.js",
-      "!src/serviceWorker.js",
-      "!src/polyfill.js"
-    ]
+  "devDependencies": {
+    "@redux-devtools/core": "^3.13.1",
+    "@redux-devtools/dock-monitor": "^3.0.1",
+    "@redux-devtools/log-monitor": "^4.0.1",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^14.2.1",
+    "auto-changelog": "~2.4.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "prettier": "2.7.1",
+    "react-scripts": "5.0.1",
+    "sass": "^1.53.0",
+    "web-vitals": "^2.1.4"
   },
   "engines": {
     "node": ">=10",
