@@ -1,5 +1,5 @@
 import React, { useState } from 'react'
-import { Link, useHistory } from 'react-router-dom'
+import { Link, useNavigate } from 'react-router-dom'
 import {
   CButton,
   CCard,
@@ -7,21 +7,22 @@ import {
   CCardGroup,
   CCol,
   CContainer,
-  CDropdown,
-  CDropdownItem,
-  CDropdownMenu,
-  CDropdownToggle,
   CForm,
-  CInput,
+  CFormInput,
   CInputGroup,
-  CInputGroupPrepend,
   CInputGroupText,
   CRow,
+  CDropdown,
+  CDropdownToggle,
+  CDropdownMenu,
+  CDropdownItem,
 } from '@coreui/react'
 import CIcon from '@coreui/icons-react'
+import { cilLockLocked, cilUser } from '@coreui/icons'
+
 import { useTranslation } from 'react-i18next'
 import { useDispatch, useSelector } from 'react-redux'
-import flagIconMap from '../../../translations/flags.json'
+import flagIconMap from '../../../translations/flags.js'
 
 // actions
 import { requestLogin } from '../../../data/actions/system/auth'
@@ -34,11 +35,11 @@ const Login = () => {
   const lang = useSelector((state) => state.ui.lang)
   const [username, setUsername] = useState('')
   const [password, setPassword] = useState('')
-  const history = useHistory()
+  const navigate = useNavigate()
 
   const handleSubmit = (e) => {
     e.preventDefault()
-    dispatch(requestLogin(username, password, history))
+    dispatch(requestLogin(username, password, navigate))
   }
 
   const selectLang = (lang) => {
@@ -46,30 +47,26 @@ const Login = () => {
   }
 
   return (
-    <div className="c-app c-default-layout flex-row align-items-center">
+    <div className="bg-light min-vh-100 d-flex flex-row align-items-center">
       <CContainer>
         <CRow className="justify-content-center">
-          <CCol md="8">
+          <CCol md={8}>
             <CCardGroup>
               <CCard className="p-4">
                 <CCardBody>
                   <CForm onSubmit={handleSubmit}>
                     <CRow>
-                      <CCol xs="6">
+                      <CCol xs={6}>
                         <h1>{t('login.login')}</h1>
-                        <p className="text-muted">
-                          {t('login.signInToYourAccount')}
-                        </p>
+                        <p className="text-muted">{t('login.signInToYourAccount')}</p>
                       </CCol>
-                      <CCol xs="6" className="text-right">
+                      <CCol xs={6} className="text-end">
                         <CDropdown className="m-1">
                           <CDropdownToggle>
-                            <CIcon name={flagIconMap[lang]} size="lg" />
+                            <CIcon icon={flagIconMap[lang]} size="lg" />
                           </CDropdownToggle>
                           <CDropdownMenu>
-                            <CDropdownItem onClick={() => selectLang('en')}>
-                              English
-                            </CDropdownItem>
+                            <CDropdownItem onClick={() => selectLang('en')}>English</CDropdownItem>
                             <CDropdownItem onClick={() => selectLang('id')}>
                               Bahasa Indonesia
                             </CDropdownItem>
@@ -78,12 +75,10 @@ const Login = () => {
                       </CCol>
                     </CRow>
                     <CInputGroup className="mb-3">
-                      <CInputGroupPrepend>
-                        <CInputGroupText>
-                          <CIcon name="cil-user" />
-                        </CInputGroupText>
-                      </CInputGroupPrepend>
-                      <CInput
+                      <CInputGroupText>
+                        <CIcon icon={cilUser} />
+                      </CInputGroupText>
+                      <CFormInput
                         type="text"
                         placeholder={t('login.username')}
                         autoComplete="username"
@@ -92,12 +87,10 @@ const Login = () => {
                       />
                     </CInputGroup>
                     <CInputGroup className="mb-4">
-                      <CInputGroupPrepend>
-                        <CInputGroupText>
-                          <CIcon name="cil-lock-locked" />
-                        </CInputGroupText>
-                      </CInputGroupPrepend>
-                      <CInput
+                      <CInputGroupText>
+                        <CIcon icon={cilLockLocked} />
+                      </CInputGroupText>
+                      <CFormInput
                         type="password"
                         placeholder={t('login.password')}
                         autoComplete="current-password"
@@ -106,7 +99,7 @@ const Login = () => {
                       />
                     </CInputGroup>
                     <CRow>
-                      <CCol xs="6">
+                      <CCol xs={6}>
                         <CButton
                           type="submit"
                           color="primary"
@@ -116,8 +109,8 @@ const Login = () => {
                           {t('login.login')}
                         </CButton>
                       </CCol>
-                      <CCol hidden xs="6" className="text-right">
-                        <CButton color="link" className="px-0">
+                      <CCol xs={6} className="text-end">
+                        <CButton color="link" className="px-0" hidden>
                           Forgot password?
                         </CButton>
                       </CCol>
@@ -125,26 +118,16 @@ const Login = () => {
                   </CForm>
                 </CCardBody>
               </CCard>
-              <CCard
-                className="text-white bg-primary py-5 d-md-down-none"
-                style={{ width: '44%' }}
-                hidden
-              >
+              <CCard className="text-white bg-primary py-5" style={{ width: '44%' }} hidden>
                 <CCardBody className="text-center">
                   <div>
                     <h2>Sign up</h2>
                     <p>
-                      Lorem ipsum dolor sit amet, consectetur adipisicing elit,
-                      sed do eiusmod tempor incididunt ut labore et dolore magna
-                      aliqua.
+                      Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod
+                      tempor incididunt ut labore et dolore magna aliqua.
                     </p>
                     <Link to="/register">
-                      <CButton
-                        color="primary"
-                        className="mt-3"
-                        active
-                        tabIndex={-1}
-                      >
+                      <CButton color="primary" className="mt-3" active tabIndex={-1}>
                         Register Now!
                       </CButton>
                     </Link>
