@@ -2,14 +2,11 @@ import { createStore, applyMiddleware, compose } from 'redux'
 import thunkMiddleware from 'redux-thunk'
 import rootReducer from '../reducers'
 import api from '../middleware/api'
-import DevTools from '../../containers/DevTools'
+import DevTools from '../../components/DevTools'
 
 const middlewares = [thunkMiddleware, api]
 
-if (
-  process.env.NODE_ENV === 'development' &&
-  process.env.REACT_APP_USE_REDUX_LOGGER === 'true'
-) {
+if (process.env.NODE_ENV === 'development' && process.env.REACT_APP_USE_REDUX_LOGGER === 'true') {
   const { createLogger } = require('redux-logger')
   const logger = createLogger()
   middlewares.push(logger)
@@ -18,7 +15,8 @@ if (
 const initialState = {
   ui: {
     lang: process.env.REACT_APP_DEFAULT_LANG,
-    sidebarShow: 'responsive',
+    sidebarUnfoldable: false,
+    sidebarShow: true,
   },
 }
 
@@ -28,7 +26,7 @@ const initStore = () => {
       ? createStore(
           rootReducer,
           initialState,
-          compose(applyMiddleware(...middlewares), DevTools.instrument())
+          compose(applyMiddleware(...middlewares), DevTools.instrument()),
         )
       : createStore(rootReducer, initialState, applyMiddleware(...middlewares))
 
