@@ -1,242 +1,195 @@
 import React from 'react'
 
-const Toaster = React.lazy(() =>
-  import('./views/notifications/toaster/Toaster')
-)
-const Tables = React.lazy(() => import('./views/base/tables/Tables'))
+const Dashboard = React.lazy(() => import('./views/dashboard/Dashboard'))
+const Colors = React.lazy(() => import('./views/theme/colors/Colors'))
+const Typography = React.lazy(() => import('./views/theme/typography/Typography'))
 
-const Breadcrumbs = React.lazy(() =>
-  import('./views/base/breadcrumbs/Breadcrumbs')
-)
+// Base
+const Accordion = React.lazy(() => import('./views/base/accordion/Accordion'))
+const Breadcrumbs = React.lazy(() => import('./views/base/breadcrumbs/Breadcrumbs'))
 const Cards = React.lazy(() => import('./views/base/cards/Cards'))
 const Carousels = React.lazy(() => import('./views/base/carousels/Carousels'))
 const Collapses = React.lazy(() => import('./views/base/collapses/Collapses'))
-const BasicForms = React.lazy(() => import('./views/base/forms/BasicForms'))
-
-const Jumbotrons = React.lazy(() =>
-  import('./views/base/jumbotrons/Jumbotrons')
-)
-const ListGroups = React.lazy(() =>
-  import('./views/base/list-groups/ListGroups')
-)
-const Navbars = React.lazy(() => import('./views/base/navbars/Navbars'))
+const ListGroups = React.lazy(() => import('./views/base/list-groups/ListGroups'))
 const Navs = React.lazy(() => import('./views/base/navs/Navs'))
-const Paginations = React.lazy(() =>
-  import('./views/base/paginations/Paginations')
-)
+const Paginations = React.lazy(() => import('./views/base/paginations/Paginations'))
+const Placeholders = React.lazy(() => import('./views/base/placeholders/Placeholders'))
 const Popovers = React.lazy(() => import('./views/base/popovers/Popovers'))
-const ProgressBar = React.lazy(() =>
-  import('./views/base/progress-bar/ProgressBar')
-)
-const Switches = React.lazy(() => import('./views/base/switches/Switches'))
-
-const Tabs = React.lazy(() => import('./views/base/tabs/Tabs'))
+const Progress = React.lazy(() => import('./views/base/progress/Progress'))
+const Spinners = React.lazy(() => import('./views/base/spinners/Spinners'))
+const Tables = React.lazy(() => import('./views/base/tables/Tables'))
 const Tooltips = React.lazy(() => import('./views/base/tooltips/Tooltips'))
-const BrandButtons = React.lazy(() =>
-  import('./views/buttons/brand-buttons/BrandButtons')
-)
-const ButtonDropdowns = React.lazy(() =>
-  import('./views/buttons/button-dropdowns/ButtonDropdowns')
-)
-const ButtonGroups = React.lazy(() =>
-  import('./views/buttons/button-groups/ButtonGroups')
-)
+
+// Buttons
 const Buttons = React.lazy(() => import('./views/buttons/buttons/Buttons'))
+const ButtonGroups = React.lazy(() => import('./views/buttons/button-groups/ButtonGroups'))
+const Dropdowns = React.lazy(() => import('./views/buttons/dropdowns/Dropdowns'))
+
+//Forms
+const ChecksRadios = React.lazy(() => import('./views/forms/checks-radios/ChecksRadios'))
+const FloatingLabels = React.lazy(() => import('./views/forms/floating-labels/FloatingLabels'))
+const FormControl = React.lazy(() => import('./views/forms/form-control/FormControl'))
+const InputGroup = React.lazy(() => import('./views/forms/input-group/InputGroup'))
+const Layout = React.lazy(() => import('./views/forms/layout/Layout'))
+const Range = React.lazy(() => import('./views/forms/range/Range'))
+const Select = React.lazy(() => import('./views/forms/select/Select'))
+const Validation = React.lazy(() => import('./views/forms/validation/Validation'))
+
 const Charts = React.lazy(() => import('./views/charts/Charts'))
-const Dashboard = React.lazy(() => import('./views/dashboard/Dashboard'))
-const CoreUIIcons = React.lazy(() =>
-  import('./views/icons/coreui-icons/CoreUIIcons')
-)
+
+// Icons
+const CoreUIIcons = React.lazy(() => import('./views/icons/coreui-icons/CoreUIIcons'))
 const Flags = React.lazy(() => import('./views/icons/flags/Flags'))
 const Brands = React.lazy(() => import('./views/icons/brands/Brands'))
+
+// Notifications
 const Alerts = React.lazy(() => import('./views/notifications/alerts/Alerts'))
 const Badges = React.lazy(() => import('./views/notifications/badges/Badges'))
 const Modals = React.lazy(() => import('./views/notifications/modals/Modals'))
-const Colors = React.lazy(() => import('./views/theme/colors/Colors'))
-const Typography = React.lazy(() =>
-  import('./views/theme/typography/Typography')
-)
-const Widgets = React.lazy(() => import('./views/widgets/Widgets'))
-const Users = React.lazy(() => import('./views/users/Users'))
-const User = React.lazy(() => import('./views/users/User'))
+const Toasts = React.lazy(() => import('./views/notifications/toasts/Toasts'))
 
-// Assets Section
-const BankAccounts = React.lazy(() =>
-  import('./views/assets/bankAccounts/BankAccounts')
-)
-const Properties = React.lazy(() =>
-  import('./views/assets/properties/Properties')
-)
+const Widgets = React.lazy(() => import('./views/widgets/Widgets'))
+
+// Assets
+const BankAccounts = React.lazy(() => import('./views/assets/bankAccounts/BankAccounts'))
+const Properties = React.lazy(() => import('./views/assets/properties/Properties'))
 const Vehicles = React.lazy(() => import('./views/assets/vehicles/Vehicles'))
 
-// Liabilities Section
-const Institutional = React.lazy(() =>
-  import('./views/liabilities/institutional/Institutional')
-)
-const Personal = React.lazy(() =>
-  import('./views/liabilities/personal/Personal')
-)
+// Liabilities
+const Institutional = React.lazy(() => import('./views/liabilities/institutional/Institutional'))
+const Personal = React.lazy(() => import('./views/liabilities/personal/Personal'))
 
-// Investments Section
-const Deposits = React.lazy(() =>
-  import('./views/investments/deposits/Deposits')
-)
-const Obligations = React.lazy(() =>
-  import('./views/investments/obligations/Obligations')
-)
+// Investments
+const Deposits = React.lazy(() => import('./views/investments/deposits/Deposits'))
+const Obligations = React.lazy(() => import('./views/investments/obligations/Obligations'))
 const Gold = React.lazy(() => import('./views/investments/gold/Gold'))
-const MutualFunds = React.lazy(() =>
-  import('./views/investments/mutualFunds/MutualFunds')
-)
+const MutualFunds = React.lazy(() => import('./views/investments/mutualFunds/MutualFunds'))
 const Stocks = React.lazy(() => import('./views/investments/stocks/Stocks'))
-const P2pLending = React.lazy(() =>
-  import('./views/investments/p2pLending/P2pLending')
-)
+const P2pLending = React.lazy(() => import('./views/investments/p2pLending/P2pLending'))
 
 const getRoutes = (t) => {
   return [
-    { path: '/', exact: true, name: t('home') },
-    { path: '/dashboard', name: 'Dashboard', component: Dashboard },
-    // Assets Section
+    { path: '/', exact: true, name: 'Home' },
+    { path: '/dashboard', name: 'Dashboard', element: Dashboard },
+    // Assets
     {
       path: '/assets',
       name: t('assets.assets'),
-      component: BankAccounts,
+      element: BankAccounts,
       exact: true,
     },
     {
       path: '/assets/bankAccounts',
       name: t('assets.bankAccounts'),
-      component: BankAccounts,
+      element: BankAccounts,
     },
     {
       path: '/assets/properties',
       name: t('assets.properties'),
-      component: Properties,
+      element: Properties,
     },
     {
       path: '/assets/vehicles',
       name: t('assets.vehicles'),
-      component: Vehicles,
+      element: Vehicles,
     },
-    // Liabilities Section
+    // Liabilities
     {
       path: '/liabilities',
       name: t('liabilities.liabilities'),
-      component: Institutional,
+      element: Institutional,
       exact: true,
     },
     {
       path: '/liabilities/institutional',
       name: t('liabilities.institutional'),
-      component: Institutional,
+      element: Institutional,
     },
     {
       path: '/liabilities/personal',
       name: t('liabilities.personal'),
-      component: Personal,
+      element: Personal,
     },
-    // Investments Section
+    // Investments
     {
       path: '/investments',
       name: t('investments.investments'),
-      component: Deposits,
+      element: Deposits,
       exact: true,
     },
     {
       path: '/investments/deposits',
       name: t('investments.deposits'),
-      component: Deposits,
+      element: Deposits,
     },
     {
       path: '/investments/obligations',
       name: t('investments.obligations'),
-      component: Obligations,
+      element: Obligations,
     },
     {
       path: '/investments/gold',
       name: t('investments.gold'),
-      component: Gold,
+      element: Gold,
     },
     {
       path: '/investments/mutualFunds',
       name: t('investments.mutualFunds'),
-      component: MutualFunds,
+      element: MutualFunds,
     },
     {
       path: '/investments/stocks',
       name: t('investments.stocks'),
-      component: Stocks,
+      element: Stocks,
     },
     {
       path: '/investments/p2pLending',
       name: t('investments.p2pLending'),
-      component: P2pLending,
+      element: P2pLending,
     },
-    // To-be-deleted Section
-    { path: '/theme', name: 'Theme', component: Colors, exact: true },
-    { path: '/theme/colors', name: 'Colors', component: Colors },
-    { path: '/theme/typography', name: 'Typography', component: Typography },
-    { path: '/base', name: 'Base', component: Cards, exact: true },
-    { path: '/base/breadcrumbs', name: 'Breadcrumbs', component: Breadcrumbs },
-    { path: '/base/cards', name: 'Cards', component: Cards },
-    { path: '/base/carousels', name: 'Carousel', component: Carousels },
-    { path: '/base/collapses', name: 'Collapse', component: Collapses },
-    { path: '/base/forms', name: 'Forms', component: BasicForms },
-    { path: '/base/jumbotrons', name: 'Jumbotrons', component: Jumbotrons },
-    { path: '/base/list-groups', name: 'List Groups', component: ListGroups },
-    { path: '/base/navbars', name: 'Navbars', component: Navbars },
-    { path: '/base/navs', name: 'Navs', component: Navs },
-    { path: '/base/paginations', name: 'Paginations', component: Paginations },
-    { path: '/base/popovers', name: 'Popovers', component: Popovers },
-    {
-      path: '/base/progress-bar',
-      name: 'Progress Bar',
-      component: ProgressBar,
-    },
-    { path: '/base/switches', name: 'Switches', component: Switches },
-    { path: '/base/tables', name: 'Tables', component: Tables },
-    { path: '/base/tabs', name: 'Tabs', component: Tabs },
-    { path: '/base/tooltips', name: 'Tooltips', component: Tooltips },
-    { path: '/buttons', name: 'Buttons', component: Buttons, exact: true },
-    { path: '/buttons/buttons', name: 'Buttons', component: Buttons },
-    {
-      path: '/buttons/button-dropdowns',
-      name: 'Dropdowns',
-      component: ButtonDropdowns,
-    },
-    {
-      path: '/buttons/button-groups',
-      name: 'Button Groups',
-      component: ButtonGroups,
-    },
-    {
-      path: '/buttons/brand-buttons',
-      name: 'Brand Buttons',
-      component: BrandButtons,
-    },
-    { path: '/charts', name: 'Charts', component: Charts },
-    { path: '/icons', exact: true, name: 'Icons', component: CoreUIIcons },
-    {
-      path: '/icons/coreui-icons',
-      name: 'CoreUI Icons',
-      component: CoreUIIcons,
-    },
-    { path: '/icons/flags', name: 'Flags', component: Flags },
-    { path: '/icons/brands', name: 'Brands', component: Brands },
-    {
-      path: '/notifications',
-      name: 'Notifications',
-      component: Alerts,
-      exact: true,
-    },
-    { path: '/notifications/alerts', name: 'Alerts', component: Alerts },
-    { path: '/notifications/badges', name: 'Badges', component: Badges },
-    { path: '/notifications/modals', name: 'Modals', component: Modals },
-    { path: '/notifications/toaster', name: 'Toaster', component: Toaster },
-    { path: '/widgets', name: 'Widgets', component: Widgets },
-    { path: '/users', exact: true, name: 'Users', component: Users },
-    { path: '/users/:id', exact: true, name: 'User Details', component: User },
+    // TODO: clean this up
+    { path: '/theme', name: 'Theme', element: Colors, exact: true },
+    { path: '/theme/colors', name: 'Colors', element: Colors },
+    { path: '/theme/typography', name: 'Typography', element: Typography },
+    { path: '/base', name: 'Base', element: Cards, exact: true },
+    { path: '/base/accordion', name: 'Accordion', element: Accordion },
+    { path: '/base/breadcrumbs', name: 'Breadcrumbs', element: Breadcrumbs },
+    { path: '/base/cards', name: 'Cards', element: Cards },
+    { path: '/base/carousels', name: 'Carousel', element: Carousels },
+    { path: '/base/collapses', name: 'Collapse', element: Collapses },
+    { path: '/base/list-groups', name: 'List Groups', element: ListGroups },
+    { path: '/base/navs', name: 'Navs', element: Navs },
+    { path: '/base/paginations', name: 'Paginations', element: Paginations },
+    { path: '/base/placeholders', name: 'Placeholders', element: Placeholders },
+    { path: '/base/popovers', name: 'Popovers', element: Popovers },
+    { path: '/base/progress', name: 'Progress', element: Progress },
+    { path: '/base/spinners', name: 'Spinners', element: Spinners },
+    { path: '/base/tables', name: 'Tables', element: Tables },
+    { path: '/base/tooltips', name: 'Tooltips', element: Tooltips },
+    { path: '/buttons', name: 'Buttons', element: Buttons, exact: true },
+    { path: '/buttons/buttons', name: 'Buttons', element: Buttons },
+    { path: '/buttons/dropdowns', name: 'Dropdowns', element: Dropdowns },
+    { path: '/buttons/button-groups', name: 'Button Groups', element: ButtonGroups },
+    { path: '/charts', name: 'Charts', element: Charts },
+    { path: '/forms', name: 'Forms', element: FormControl, exact: true },
+    { path: '/forms/form-control', name: 'Form Control', element: FormControl },
+    { path: '/forms/select', name: 'Select', element: Select },
+    { path: '/forms/checks-radios', name: 'Checks & Radios', element: ChecksRadios },
+    { path: '/forms/range', name: 'Range', element: Range },
+    { path: '/forms/input-group', name: 'Input Group', element: InputGroup },
+    { path: '/forms/floating-labels', name: 'Floating Labels', element: FloatingLabels },
+    { path: '/forms/layout', name: 'Layout', element: Layout },
+    { path: '/forms/validation', name: 'Validation', element: Validation },
+    { path: '/icons', exact: true, name: 'Icons', element: CoreUIIcons },
+    { path: '/icons/coreui-icons', name: 'CoreUI Icons', element: CoreUIIcons },
+    { path: '/icons/flags', name: 'Flags', element: Flags },
+    { path: '/icons/brands', name: 'Brands', element: Brands },
+    { path: '/notifications', name: 'Notifications', element: Alerts, exact: true },
+    { path: '/notifications/alerts', name: 'Alerts', element: Alerts },
+    { path: '/notifications/badges', name: 'Badges', element: Badges },
+    { path: '/notifications/modals', name: 'Modals', element: Modals },
+    { path: '/notifications/toasts', name: 'Toasts', element: Toasts },
+    { path: '/widgets', name: 'Widgets', element: Widgets },
   ]
 }
 
