@@ -2,71 +2,68 @@ import React from 'react'
 import {
   CBreadcrumb,
   CBreadcrumbItem,
-  CBreadcrumbRouter,
   CCard,
   CCardBody,
   CCardHeader,
-  CLink,
   CCol,
   CRow,
+  CLink,
 } from '@coreui/react'
-import { useTranslation } from 'react-i18next'
-
-import getRoutes from '../../../routes'
+import { DocsExample } from 'src/components'
 
 const Breadcrumbs = () => {
-  const { t } = useTranslation('navigation')
-  const routes = getRoutes(t)
   return (
     <CRow>
-      <CCol xs="12">
-        <CCard>
+      <CCol xs={12}>
+        <CCard className="mb-4">
           <CCardHeader>
-            Bootstrap Breadcrumb
-            <div className="card-header-actions">
-              <a
-                href="https://coreui.github.io/components/breadcrumbs/"
-                rel="noreferrer noopener"
-                target="_blank"
-                className="card-header-action"
-              >
-                <small className="text-muted">docs</small>
-              </a>
-            </div>
+            <strong>React Breadcrumb</strong>
           </CCardHeader>
           <CCardBody>
-            <h6>CBreadcrumbRouter wrapper component</h6>
-            <CBreadcrumbRouter routes={routes} />
-            <h6>Manual</h6>
-            <CBreadcrumb>
-              <CBreadcrumbItem>
-                <CLink>Home</CLink>
-              </CBreadcrumbItem>
-              <CBreadcrumbItem active>Library</CBreadcrumbItem>
-            </CBreadcrumb>
-            <CBreadcrumb>
-              <CBreadcrumbItem>
-                <CLink>Home</CLink>
-              </CBreadcrumbItem>
-              <CBreadcrumbItem>
-                <CLink>Library</CLink>
-              </CBreadcrumbItem>
-              <CBreadcrumbItem active>Data</CBreadcrumbItem>
-            </CBreadcrumb>
-            <CBreadcrumb>
-              <CBreadcrumbItem>
-                <CLink>Home</CLink>
-              </CBreadcrumbItem>
-              <CBreadcrumbItem>
-                <CLink>Library</CLink>
-              </CBreadcrumbItem>
-              <CBreadcrumbItem>
-                <CLink>Data</CLink>
-              </CBreadcrumbItem>
-              <CBreadcrumbItem active>
-                <span>Bootstrap</span>
-              </CBreadcrumbItem>
-            </CBreadcrumb>
+            <p className="text-medium-emphasis small">
+              The breadcrumb navigation provides links back to each previous page the user navigated
+              through and shows the current location in a website or an application. You don’t have
+              to add separators, because they automatically added in CSS through{' '}
+              <a href="https://developer.mozilla.org/en-US/docs/Web/CSS/::before">
+                {' '}
+                <code>::before</code>
+              </a>{' '}
+              and{' '}
+              <a href="https://developer.mozilla.org/en-US/docs/Web/CSS/content">
+                {' '}
+                <code>content</code>
+              </a>
+              .
+            </p>
+            <DocsExample href="components/breadcrumb">
+              <CBreadcrumb>
+                <CBreadcrumbItem>
+                  <CLink href="#">Home</CLink>
+                </CBreadcrumbItem>
+                <CBreadcrumbItem active>Library</CBreadcrumbItem>
+              </CBreadcrumb>
+              <CBreadcrumb>
+                <CBreadcrumbItem>
+                  <CLink href="#">Home</CLink>
+                </CBreadcrumbItem>
+                <CBreadcrumbItem>
+                  <CLink href="#">Library</CLink>
+                </CBreadcrumbItem>
+                <CBreadcrumbItem active>Data</CBreadcrumbItem>
+              </CBreadcrumb>
+              <CBreadcrumb>
+                <CBreadcrumbItem>
+                  <CLink href="#">Home</CLink>
+                </CBreadcrumbItem>
+                <CBreadcrumbItem>
+                  <CLink href="#">Library</CLink>
+                </CBreadcrumbItem>
+                <CBreadcrumbItem>
+                  <CLink href="#">Data</CLink>
+                </CBreadcrumbItem>
+                <CBreadcrumbItem active>Bootstrap</CBreadcrumbItem>
+              </CBreadcrumb>
+            </DocsExample>
           </CCardBody>
         </CCard>
       </CCol>
