@@ -1,203 +1,207 @@
-import React, { useState } from 'react'
+import React from 'react'
 import {
   CCard,
   CCardBody,
   CCardHeader,
   CCarousel,
   CCarouselCaption,
-  CCarouselControl,
-  CCarouselIndicators,
-  CCarouselInner,
   CCarouselItem,
   CCol,
   CRow,
 } from '@coreui/react'
+import { DocsExample } from 'src/components'
 
-const slides = [
-  'data:image/svg+xml;charset=UTF-8,%3Csvg%20width%3D%22800%22%20height%3D%22400%22%20xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22%20viewBox%3D%220%200%20800%20400%22%20preserveAspectRatio%3D%22none%22%3E%3Cdefs%3E%3Cstyle%20type%3D%22text%2Fcss%22%3E%23holder_1607923e7e2%20text%20%7B%20fill%3A%23555%3Bfont-weight%3Anormal%3Bfont-family%3AHelvetica%2C%20monospace%3Bfont-size%3A40pt%20%7D%20%3C%2Fstyle%3E%3C%2Fdefs%3E%3Cg%20id%3D%22holder_1607923e7e2%22%3E%3Crect%20width%3D%22800%22%20height%3D%22400%22%20fill%3D%22%23777%22%3E%3C%2Frect%3E%3Cg%3E%3Ctext%20x%3D%22285.9296875%22%20y%3D%22217.75625%22%3EFirst%20slide%3C%2Ftext%3E%3C%2Fg%3E%3C%2Fg%3E%3C%2Fsvg%3E',
-  'data:image/svg+xml;charset=UTF-8,%3Csvg%20width%3D%22800%22%20height%3D%22400%22%20xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22%20viewBox%3D%220%200%20800%20400%22%20preserveAspectRatio%3D%22none%22%3E%3Cdefs%3E%3Cstyle%20type%3D%22text%2Fcss%22%3E%23holder_15ba800aa20%20text%20%7B%20fill%3A%23444%3Bfont-weight%3Anormal%3Bfont-family%3AHelvetica%2C%20monospace%3Bfont-size%3A40pt%20%7D%20%3C%2Fstyle%3E%3C%2Fdefs%3E%3Cg%20id%3D%22holder_15ba800aa20%22%3E%3Crect%20width%3D%22800%22%20height%3D%22400%22%20fill%3D%22%23666%22%3E%3C%2Frect%3E%3Cg%3E%3Ctext%20x%3D%22247.3203125%22%20y%3D%22218.3%22%3ESecond%20slide%3C%2Ftext%3E%3C%2Fg%3E%3C%2Fg%3E%3C%2Fsvg%3E',
-  'data:image/svg+xml;charset=UTF-8,%3Csvg%20width%3D%22800%22%20height%3D%22400%22%20xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22%20viewBox%3D%220%200%20800%20400%22%20preserveAspectRatio%3D%22none%22%3E%3Cdefs%3E%3Cstyle%20type%3D%22text%2Fcss%22%3E%23holder_15ba800aa21%20text%20%7B%20fill%3A%23333%3Bfont-weight%3Anormal%3Bfont-family%3AHelvetica%2C%20monospace%3Bfont-size%3A40pt%20%7D%20%3C%2Fstyle%3E%3C%2Fdefs%3E%3Cg%20id%3D%22holder_15ba800aa21%22%3E%3Crect%20width%3D%22800%22%20height%3D%22400%22%20fill%3D%22%23555%22%3E%3C%2Frect%3E%3Cg%3E%3Ctext%20x%3D%22277%22%20y%3D%22218.3%22%3EThird%20slide%3C%2Ftext%3E%3C%2Fg%3E%3C%2Fg%3E%3C%2Fsvg%3E',
+import AngularImg from 'src/assets/images/angular.jpg'
+import ReactImg from 'src/assets/images/react.jpg'
+import VueImg from 'src/assets/images/vue.jpg'
+
+const slidesLight = [
+  'data:image/svg+xml;charset=UTF-8,%3Csvg%20width%3D%22800%22%20height%3D%22400%22%20xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22%20viewBox%3D%220%200%20800%20400%22%20preserveAspectRatio%3D%22none%22%3E%3Cdefs%3E%3Cstyle%20type%3D%22text%2Fcss%22%3E%23holder_1607923e7e2%20text%20%7B%20fill%3A%23AAA%3Bfont-weight%3Anormal%3Bfont-family%3AHelvetica%2C%20monospace%3Bfont-size%3A40pt%20%7D%20%3C%2Fstyle%3E%3C%2Fdefs%3E%3Cg%20id%3D%22holder_1607923e7e2%22%3E%3Crect%20width%3D%22800%22%20height%3D%22400%22%20fill%3D%22%23F5F5F5%22%3E%3C%2Frect%3E%3Cg%3E%3Ctext%20x%3D%22285.9296875%22%20y%3D%22217.75625%22%3EFirst%20slide%3C%2Ftext%3E%3C%2Fg%3E%3C%2Fg%3E%3C%2Fsvg%3E',
+  'data:image/svg+xml;charset=UTF-8,%3Csvg%20width%3D%22800%22%20height%3D%22400%22%20xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22%20viewBox%3D%220%200%20800%20400%22%20preserveAspectRatio%3D%22none%22%3E%3Cdefs%3E%3Cstyle%20type%3D%22text%2Fcss%22%3E%23holder_15ba800aa20%20text%20%7B%20fill%3A%23BBB%3Bfont-weight%3Anormal%3Bfont-family%3AHelvetica%2C%20monospace%3Bfont-size%3A40pt%20%7D%20%3C%2Fstyle%3E%3C%2Fdefs%3E%3Cg%20id%3D%22holder_15ba800aa20%22%3E%3Crect%20width%3D%22800%22%20height%3D%22400%22%20fill%3D%22%23EEE%22%3E%3C%2Frect%3E%3Cg%3E%3Ctext%20x%3D%22247.3203125%22%20y%3D%22218.3%22%3ESecond%20slide%3C%2Ftext%3E%3C%2Fg%3E%3C%2Fg%3E%3C%2Fsvg%3E',
+  'data:image/svg+xml;charset=UTF-8,%3Csvg%20width%3D%22800%22%20height%3D%22400%22%20xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22%20viewBox%3D%220%200%20800%20400%22%20preserveAspectRatio%3D%22none%22%3E%3Cdefs%3E%3Cstyle%20type%3D%22text%2Fcss%22%3E%23holder_15ba800aa21%20text%20%7B%20fill%3A%23999%3Bfont-weight%3Anormal%3Bfont-family%3AHelvetica%2C%20monospace%3Bfont-size%3A40pt%20%7D%20%3C%2Fstyle%3E%3C%2Fdefs%3E%3Cg%20id%3D%22holder_15ba800aa21%22%3E%3Crect%20width%3D%22800%22%20height%3D%22400%22%20fill%3D%22%23E5E5E5%22%3E%3C%2Frect%3E%3Cg%3E%3Ctext%20x%3D%22277%22%20y%3D%22218.3%22%3EThird%20slide%3C%2Ftext%3E%3C%2Fg%3E%3C%2Fg%3E%3C%2Fsvg%3E',
 ]
 
 const Carousels = () => {
-  const [activeIndex] = useState(1)
-
   return (
     <CRow>
-      <CCol xs="12" xl="6">
-        <CCard>
-          <CCardHeader>Carousel with controls</CCardHeader>
-          <CCardBody>
-            <CCarousel>
-              <CCarouselInner>
-                <CCarouselItem>
-                  <img
-                    className="d-block w-100"
-                    src={slides[0]}
-                    alt="slide 1"
-                  />
-                </CCarouselItem>
-                <CCarouselItem>
-                  <img
-                    className="d-block w-100"
-                    src={slides[1]}
-                    alt="slide 2"
-                  />
-                </CCarouselItem>
-                <CCarouselItem>
-                  <img
-                    className="d-block w-100"
-                    src={slides[2]}
-                    alt="slide 3"
-                  />
-                </CCarouselItem>
-              </CCarouselInner>
-              <CCarouselControl direction="prev" />
-              <CCarouselControl direction="next" />
-            </CCarousel>
-          </CCardBody>
-        </CCard>
-      </CCol>
-      <CCol xs="12" xl="6">
-        <CCard>
+      <CCol xs={12}>
+        <CCard className="mb-4">
           <CCardHeader>
-            Carousel with controls, indicators and caption
+            <strong>Carousel</strong> <small>Slide only</small>
           </CCardHeader>
           <CCardBody>
-            <CCarousel activeIndex={activeIndex}>
-              <CCarouselIndicators />
-              <CCarouselInner>
+            <p className="text-medium-emphasis small">Here’s a carousel with slides</p>
+            <DocsExample href="components/carousel">
+              <CCarousel>
                 <CCarouselItem>
-                  <img
-                    className="d-block w-100"
-                    src={slides[0]}
-                    alt="slide 1"
-                  />
-                  <CCarouselCaption>
-                    <h3>Slide 1</h3>
-                    <p>Slide 1</p>
-                  </CCarouselCaption>
+                  <img className="d-block w-100" src={ReactImg} alt="slide 1" />
                 </CCarouselItem>
                 <CCarouselItem>
-                  <img
-                    className="d-block w-100"
-                    src={slides[1]}
-                    alt="slide 2"
-                  />
-                  <CCarouselCaption>
-                    <h3>Slide 2</h3>
-                    <p>Slide 2</p>
-                  </CCarouselCaption>
+                  <img className="d-block w-100" src={AngularImg} alt="slide 2" />
                 </CCarouselItem>
                 <CCarouselItem>
-                  <img
-                    className="d-block w-100"
-                    src={slides[2]}
-                    alt="slide 3"
-                  />
-                  <CCarouselCaption>
-                    <h3>Slide 3</h3>
-                    <p>Slide 3</p>
-                  </CCarouselCaption>
+                  <img className="d-block w-100" src={VueImg} alt="slide 3" />
                 </CCarouselItem>
-              </CCarouselInner>
-              <CCarouselControl direction="prev" />
-              <CCarouselControl direction="next" />
-            </CCarousel>
+              </CCarousel>
+            </DocsExample>
           </CCardBody>
         </CCard>
       </CCol>
-      <CCol xs="12" xl="6">
-        <CCard>
-          <CCardHeader>Carousel animation</CCardHeader>
+      <CCol xs={12}>
+        <CCard className="mb-4">
+          <CCardHeader>
+            <strong>Carousel</strong> <small>With controls</small>
+          </CCardHeader>
           <CCardBody>
-            <CCarousel animate>
-              <CCarouselIndicators />
-              <CCarouselInner>
+            <p className="text-medium-emphasis small">
+              Adding in the previous and next controls by <code>controls</code> property.
+            </p>
+            <DocsExample href="components/carousel/#with-controls">
+              <CCarousel controls>
                 <CCarouselItem>
-                  <img
-                    className="d-block w-100"
-                    src={slides[0]}
-                    alt="slide 1"
-                  />
-                  <CCarouselCaption>
-                    <h3>Slide 1</h3>
-                    <p>Slide 1</p>
-                  </CCarouselCaption>
+                  <img className="d-block w-100" src={ReactImg} alt="slide 1" />
                 </CCarouselItem>
                 <CCarouselItem>
-                  <img
-                    className="d-block w-100"
-                    src={slides[1]}
-                    alt="slide 2"
-                  />
-                  <CCarouselCaption>
-                    <h3>Slide 2</h3>
-                    <p>Slide 2</p>
-                  </CCarouselCaption>
+                  <img className="d-block w-100" src={AngularImg} alt="slide 2" />
                 </CCarouselItem>
                 <CCarouselItem>
-                  <img
-                    className="d-block w-100"
-                    src={slides[2]}
-                    alt="slide 3"
-                  />
-                  <CCarouselCaption>
-                    <h3>Slide 3</h3>
-                    <p>Slide 3</p>
-                  </CCarouselCaption>
+                  <img className="d-block w-100" src={VueImg} alt="slide 3" />
                 </CCarouselItem>
-              </CCarouselInner>
-              <CCarouselControl direction="prev" />
-              <CCarouselControl direction="next" />
-            </CCarousel>
+              </CCarousel>
+            </DocsExample>
           </CCardBody>
         </CCard>
       </CCol>
-      <CCol xs="12" xl="6">
-        <CCard>
-          <CCardHeader>Carousel animation with autoSlide</CCardHeader>
+      <CCol xs={12}>
+        <CCard className="mb-4">
+          <CCardHeader>
+            <strong>Carousel</strong> <small>With indicators</small>
+          </CCardHeader>
           <CCardBody>
-            <CCarousel animate autoSlide={3000}>
-              <CCarouselIndicators />
-              <CCarouselInner>
+            <p className="text-medium-emphasis small">
+              You can attach the indicators to the carousel, lengthwise the controls, too.
+            </p>
+            <DocsExample href="components/carousel/#with-indicators">
+              <CCarousel controls indicators>
                 <CCarouselItem>
-                  <img
-                    className="d-block w-100"
-                    src={slides[0]}
-                    alt="slide 1"
-                  />
-                  <CCarouselCaption>
-                    <h3>Slide 1</h3>
-                    <p>Slide 1</p>
+                  <img className="d-block w-100" src={ReactImg} alt="slide 1" />
+                </CCarouselItem>
+                <CCarouselItem>
+                  <img className="d-block w-100" src={AngularImg} alt="slide 2" />
+                </CCarouselItem>
+                <CCarouselItem>
+                  <img className="d-block w-100" src={VueImg} alt="slide 3" />
+                </CCarouselItem>
+              </CCarousel>
+            </DocsExample>
+          </CCardBody>
+        </CCard>
+      </CCol>
+      <CCol xs={12}>
+        <CCard className="mb-4">
+          <CCardHeader>
+            <strong>Carousel</strong> <small>With captions</small>
+          </CCardHeader>
+          <CCardBody>
+            <p className="text-medium-emphasis small">
+              You can add captions to slides with the <code>&lt;CCarouselCaption&gt;</code> element
+              within any <code>&lt;CCarouselItem&gt;</code>. They can be immediately hidden on
+              smaller viewports, as shown below, with optional{' '}
+              <a href="https://coreui.io/4.0/utilities/display">display utilities</a>. We hide them
+              with <code>.d-none</code> and draw them back on medium-sized devices with{' '}
+              <code>.d-md-block</code>.
+            </p>
+            <DocsExample href="components/carousel/#with-captions">
+              <CCarousel controls indicators>
+                <CCarouselItem>
+                  <img className="d-block w-100" src={ReactImg} alt="slide 1" />
+                  <CCarouselCaption className="d-none d-md-block">
+                    <h5>First slide label</h5>
+                    <p>Some representative placeholder content for the first slide.</p>
                   </CCarouselCaption>
                 </CCarouselItem>
                 <CCarouselItem>
-                  <img
-                    className="d-block w-100"
-                    src={slides[1]}
-                    alt="slide 2"
-                  />
-                  <CCarouselCaption>
-                    <h3>Slide 2</h3>
-                    <p>Slide 2</p>
+                  <img className="d-block w-100" src={AngularImg} alt="slide 2" />
+                  <CCarouselCaption className="d-none d-md-block">
+                    <h5>Second slide label</h5>
+                    <p>Some representative placeholder content for the first slide.</p>
                   </CCarouselCaption>
                 </CCarouselItem>
                 <CCarouselItem>
-                  <img
-                    className="d-block w-100"
-                    src={slides[2]}
-                    alt="slide 3"
-                  />
-                  <CCarouselCaption>
-                    <h3>Slide 3</h3>
-                    <p>Slide 3</p>
+                  <img className="d-block w-100" src={VueImg} alt="slide 3" />
+                  <CCarouselCaption className="d-none d-md-block">
+                    <h5>Third slide label</h5>
+                    <p>Some representative placeholder content for the first slide.</p>
                   </CCarouselCaption>
                 </CCarouselItem>
-              </CCarouselInner>
-              <CCarouselControl direction="prev" />
-              <CCarouselControl direction="next" />
-            </CCarousel>
+              </CCarousel>
+            </DocsExample>
+          </CCardBody>
+        </CCard>
+      </CCol>
+      <CCol xs={12}>
+        <CCard className="mb-4">
+          <CCardHeader>
+            <strong>Carousel</strong> <small>Crossfade</small>
+          </CCardHeader>
+          <CCardBody>
+            <p className="text-medium-emphasis small">
+              Add <code>transition=&#34;crossfade&#34;</code> to your carousel to animate slides
+              with a fade transition instead of a slide.
+            </p>
+            <DocsExample href="components/carousel/#crossfade">
+              <CCarousel controls transition="crossfade">
+                <CCarouselItem>
+                  <img className="d-block w-100" src={ReactImg} alt="slide 1" />
+                </CCarouselItem>
+                <CCarouselItem>
+                  <img className="d-block w-100" src={AngularImg} alt="slide 2" />
+                </CCarouselItem>
+                <CCarouselItem>
+                  <img className="d-block w-100" src={VueImg} alt="slide 3" />
+                </CCarouselItem>
+              </CCarousel>
+            </DocsExample>
+          </CCardBody>
+        </CCard>
+      </CCol>
+      <CCol xs={12}>
+        <CCard className="mb-4">
+          <CCardHeader>
+            <strong>Carousel</strong> <small>Dark variant</small>
+          </CCardHeader>
+          <CCardBody>
+            <p className="text-medium-emphasis small">
+              Add <code>dark</code> property to the <code>CCarousel</code> for darker controls,
+              indicators, and captions. Controls have been inverted from their default white fill
+              with the <code>filter</code> CSS property. Captions and controls have additional Sass
+              variables that customize the <code>color</code> and <code>background-color</code>.
+            </p>
+            <DocsExample href="components/carousel/#dark-variant">
+              <CCarousel controls indicators dark>
+                <CCarouselItem>
+                  <img className="d-block w-100" src={slidesLight[0]} alt="slide 1" />
+                  <CCarouselCaption className="d-none d-md-block">
+                    <h5>First slide label</h5>
+                    <p>Some representative placeholder content for the first slide.</p>
+                  </CCarouselCaption>
+                </CCarouselItem>
+                <CCarouselItem>
+                  <img className="d-block w-100" src={slidesLight[1]} alt="slide 2" />
+                  <CCarouselCaption className="d-none d-md-block">
+                    <h5>Second slide label</h5>
+                    <p>Some representative placeholder content for the first slide.</p>
+                  </CCarouselCaption>
+                </CCarouselItem>
+                <CCarouselItem>
+                  <img className="d-block w-100" src={slidesLight[2]} alt="slide 3" />
+                  <CCarouselCaption className="d-none d-md-block">
+                    <h5>Third slide label</h5>
+                    <p>Some representative placeholder content for the first slide.</p>
+                  </CCarouselCaption>
+                </CCarouselItem>
+              </CCarousel>
+            </DocsExample>
           </CCardBody>
         </CCard>
       </CCol>
