@@ -1,90 +1,54 @@
 import React from 'react'
-import {
-  CButton,
-  CCard,
-  CCardBody,
-  CCardHeader,
-  CCol,
-  CContainer,
-  CJumbotron,
-  CRow,
-  CEmbed,
-  CEmbedItem,
-} from '@coreui/react'
+import { CButton, CCard, CCardBody, CCardHeader, CCol, CContainer, CRow } from '@coreui/react'
+import { DocsLink } from 'src/components'
 
 const Jumbotrons = () => {
   return (
     <>
-      <CRow>
-        <CCol>
-          <CCard>
-            <CCardHeader>
-              Jumbotron
-              <div className="card-header-actions">
-                <a
-                  href="https://coreui.github.io/components/jumbotron/"
-                  rel="noreferrer noopener"
-                  target="_blank"
-                  className="card-header-action"
-                >
-                  <small className="text-muted">docs</small>
-                </a>
-              </div>
-            </CCardHeader>
-            <CCardBody>
-              <CJumbotron className="border">
-                <h1 className="display-3">Hello, world!</h1>
-                <p className="lead">
-                  This is a simple hero unit, a simple Jumbotron - style
-                  component for calling extra attention to featured content or
-                  information.
-                </p>
-                <hr className="my-2" />
+      <CCard className="mb-4">
+        <CCardHeader>
+          Jumbotron
+          <DocsLink name="CJumbotron" />
+        </CCardHeader>
+        <CCardBody>
+          <CContainer className="py-5" fluid>
+            <h1 className="display-5 fw-bold">Custom jumbotron</h1>
+            <p className="col-md-8 fs-4">
+              Using a series of utilities, you can create this jumbotron, just like the one in
+              previous versions of Bootstrap. Check out the examples below for how you can remix and
+              restyle it to your liking.
+            </p>
+            <CButton size="lg">Example button</CButton>
+          </CContainer>
+          <CRow className="align-items-md-stretch">
+            <CCol md={6}>
+              <div className="h-100 p-5 text-white bg-dark rounded-3">
+                <h2>Change the background</h2>
                 <p>
-                  It uses utility classes for typgraphy and spacing to space
-                  content out within the larger container.
+                  Swap the background-color utility and add a `.text-*` color utility to mix up the
+                  jumbotron look. Then, mix and match with additional component themes and more.
                 </p>
-                <p className="lead">
-                  <CButton color="primary" size="lg">
-                    Learn More
-                  </CButton>
+                <CButton color="light" variant="outline">
+                  DocsExample button
+                </CButton>
+              </div>
+            </CCol>
+            <CCol md={6}>
+              <div className="h-100 p-5 bg-light border rounded-3">
+                <h2>Add borders</h2>
+                <p>
+                  Or, keep it light and add a border for some added definition to the boundaries of
+                  your content. Be sure to look under the hood at the source HTML here as we&#39;ve
+                  adjusted the alignment and sizing of both column&#39;s content for equal-height.
                 </p>
-              </CJumbotron>
-            </CCardBody>
-          </CCard>
-        </CCol>
-        <CCol>
-          <CCard>
-            <CCardHeader>
-              Jumbotron
-              <small> fluid</small>
-            </CCardHeader>
-            <CCardBody>
-              <CJumbotron fluid>
-                <CContainer fluid>
-                  <h1 className="display-3">Fluid jumbotron</h1>
-                  <p className="lead">
-                    This is a modified jumbotron that occupies the entire
-                    horizontal space of its parent.
-                  </p>
-                </CContainer>
-              </CJumbotron>
-            </CCardBody>
-          </CCard>
-        </CCol>
-      </CRow>
-      <CRow>
-        <CCol>
-          <CCard>
-            <CCardHeader>Embed</CCardHeader>
-            <CCardBody>
-              <CEmbed>
-                <CEmbedItem src="https://www.youtube.com/embed/zpOULjyy-n8?rel=0" />
-              </CEmbed>
-            </CCardBody>
-          </CCard>
-        </CCol>
-      </CRow>
+                <CButton color="secondary" variant="outline">
+                  DocsExample button
+                </CButton>
+              </div>
+            </CCol>
+          </CRow>
+        </CCardBody>
+      </CCard>
     </>
   )
 }
