@@ -1,1182 +1,400 @@
 import React from 'react'
-import {
-  CButton,
-  CCard,
-  CCardBody,
-  CCardHeader,
-  CCol,
-  CRow,
-} from '@coreui/react'
+import { CButton, CCard, CCardBody, CCardHeader, CCol, CRow } from '@coreui/react'
 import CIcon from '@coreui/icons-react'
+import { cilBell } from '@coreui/icons'
+import { DocsExample } from 'src/components'
 
 const Buttons = () => {
   return (
-    <>
-      <CCard>
-        <CCardHeader>Standard Buttons</CCardHeader>
-        <CCardBody>
-          <CRow className="align-items-center">
-            <CCol col="12" xl className="mb-3 mb-xl-0">
-              Normal
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block color="primary">
-                Primary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block color="secondary">
-                Secondary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block color="success">
-                Success
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block color="warning">
-                Warning
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block color="danger">
-                Danger
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block color="info">
-                Info
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block color="light">
-                Light
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block color="dark">
-                Dark
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block color="link">
+    <CRow>
+      <CCol xs={12}>
+        <CCard className="mb-4">
+          <CCardHeader>
+            <strong>React Button</strong>
+          </CCardHeader>
+          <CCardBody>
+            <p className="text-medium-emphasis small">
+              CoreUI includes a bunch of predefined buttons components, each serving its own
+              semantic purpose. Buttons show what action will happen when the user clicks or touches
+              it. CoreUI buttons are used to initialize operations, both in the background or
+              foreground of an experience.
+            </p>
+            <DocsExample href="components/buttons">
+              {['normal', 'active', 'disabled'].map((state, index) => (
+                <CRow className="align-items-center mb-3" key={index}>
+                  <CCol xs={12} xl={2} className="mb-3 mb-xl-0">
+                    {state.charAt(0).toUpperCase() + state.slice(1)}
+                  </CCol>
+                  <CCol xs>
+                    {[
+                      'primary',
+                      'secondary',
+                      'success',
+                      'danger',
+                      'warning',
+                      'info',
+                      'light',
+                      'dark',
+                    ].map((color, index) => (
+                      <CButton
+                        color={color}
+                        key={index}
+                        active={state === 'active'}
+                        disabled={state === 'disabled'}
+                      >
+                        {color.charAt(0).toUpperCase() + color.slice(1)}
+                      </CButton>
+                    ))}
+                    <CButton color="link">Link</CButton>
+                  </CCol>
+                </CRow>
+              ))}
+            </DocsExample>
+          </CCardBody>
+        </CCard>
+      </CCol>
+      <CCol xs={12}>
+        <CCard className="mb-4">
+          <CCardHeader>
+            <strong>React Button</strong> <small>with icons</small>
+          </CCardHeader>
+          <CCardBody>
+            <p className="text-medium-emphasis small">
+              You can combine button with our <a href="https://icons.coreui.io/">CoreUI Icons</a>.
+            </p>
+            <DocsExample href="components/buttons">
+              {['normal', 'active', 'disabled'].map((state, index) => (
+                <CRow className="align-items-center mb-3" key={index}>
+                  <CCol xs={12} xl={2} className="mb-3 mb-xl-0">
+                    {state.charAt(0).toUpperCase() + state.slice(1)}
+                  </CCol>
+                  <CCol xs>
+                    {[
+                      'primary',
+                      'secondary',
+                      'success',
+                      'danger',
+                      'warning',
+                      'info',
+                      'light',
+                      'dark',
+                    ].map((color, index) => (
+                      <CButton
+                        color={color}
+                        key={index}
+                        active={state === 'active'}
+                        disabled={state === 'disabled'}
+                      >
+                        <CIcon icon={cilBell} className="me-2" />
+                        {color.charAt(0).toUpperCase() + color.slice(1)}
+                      </CButton>
+                    ))}
+                    <CButton color="link">
+                      <CIcon icon={cilBell} className="me-2" />
+                      Link
+                    </CButton>
+                  </CCol>
+                </CRow>
+              ))}
+            </DocsExample>
+          </CCardBody>
+        </CCard>
+      </CCol>
+      <CCol xs={12}>
+        <CCard className="mb-4">
+          <CCardHeader>
+            <strong>React Button</strong> <small>Button components</small>
+          </CCardHeader>
+          <CCardBody>
+            <p className="text-medium-emphasis small">
+              The <code>&lt;CButton&gt;</code> component are designed for{' '}
+              <code>&lt;button&gt;</code> , <code>&lt;a&gt;</code> or <code>&lt;input&gt;</code>{' '}
+              elements (though some browsers may apply a slightly different rendering).
+            </p>
+            <p className="text-medium-emphasis small">
+              If you&#39;re using <code>&lt;CButton&gt;</code> component as <code>&lt;a&gt;</code>{' '}
+              elements that are used to trigger functionality ex. collapsing content, these links
+              should be given a <code>role=&#34;button&#34;</code> to adequately communicate their
+              meaning to assistive technologies such as screen readers.
+            </p>
+            <DocsExample href="components/buttons#button-components">
+              <CButton component="a" color="primary" href="#" role="button">
                 Link
               </CButton>
-            </CCol>
-          </CRow>
-          <CRow className="align-items-center mt-3">
-            <CCol col="12" xl className="mb-3 mb-xl-0">
-              Active State
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton active block color="primary" aria-pressed="true">
-                Primary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton active block color="secondary" aria-pressed="true">
-                Secondary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton active block color="success" aria-pressed="true">
-                Success
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton active block color="warning" aria-pressed="true">
-                Warning
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton active block color="danger" aria-pressed="true">
-                Danger
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton active block color="info" aria-pressed="true">
-                Info
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton active block color="light" aria-pressed="true">
-                Light
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton active block color="dark" aria-pressed="true">
-                Dark
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton active block color="link" aria-pressed="true">
-                Link
-              </CButton>
-            </CCol>
-          </CRow>
-          <CRow className="align-items-center mt-3">
-            <CCol col="12" xl className="mb-3 mb-xl-0">
-              Disabled
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block color="primary" disabled>
-                Primary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block color="secondary" disabled>
-                Secondary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block color="success" disabled>
-                Success
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block color="warning" disabled>
-                Warning
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block color="danger" disabled>
-                Danger
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block color="info" disabled>
-                Info
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block color="light" disabled>
-                Light
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block color="dark" disabled>
-                Dark
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block color="link" disabled>
-                Link
-              </CButton>
-            </CCol>
-          </CRow>
-        </CCardBody>
-      </CCard>
-      <CCard>
-        <CCardHeader>variant="outline" Buttons</CCardHeader>
-        <CCardBody>
-          <p>
-            With <code>outline</code> prop.
-          </p>
-          <CRow className="align-items-center">
-            <CCol col="12" xl className="mb-3 mb-xl-0">
-              Normal
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="outline" color="primary">
-                Primary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="outline" color="secondary">
-                Secondary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="outline" color="success">
-                Success
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="outline" color="warning">
-                Warning
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="outline" color="danger">
-                Danger
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="outline" color="info">
-                Info
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="outline" color="light">
-                Light
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="outline" color="dark">
-                Dark
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0"></CCol>
-          </CRow>
-          <CRow className="align-items-center mt-3">
-            <CCol col="12" xl className="mb-3 mb-xl-0">
-              Active State
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                block
-                variant="outline"
-                active
-                color="primary"
-                aria-pressed="true"
-              >
-                Primary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                block
-                variant="outline"
-                active
-                color="secondary"
-                aria-pressed="true"
-              >
-                Secondary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                block
-                variant="outline"
-                active
-                color="success"
-                aria-pressed="true"
-              >
-                Success
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                block
-                variant="outline"
-                active
-                color="warning"
-                aria-pressed="true"
-              >
-                Warning
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                block
-                variant="outline"
-                active
-                color="danger"
-                aria-pressed="true"
-              >
-                Danger
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                block
-                variant="outline"
-                active
-                color="info"
-                aria-pressed="true"
-              >
-                Info
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                block
-                variant="outline"
-                active
-                color="light"
-                aria-pressed="true"
-              >
-                Light
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                block
-                variant="outline"
-                active
-                color="dark"
-                aria-pressed="true"
-              >
-                Dark
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0"></CCol>
-          </CRow>
-          <CRow className="align-items-center mt-3">
-            <CCol col="12" xl className="mb-3 mb-xl-0">
-              Disabled
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="outline" color="primary" disabled>
-                Primary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="outline" color="secondary" disabled>
-                Secondary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="outline" color="success" disabled>
-                Success
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="outline" color="warning" disabled>
-                Warning
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="outline" color="danger" disabled>
-                Danger
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="outline" color="info" disabled>
-                Info
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="outline" color="light" disabled>
-                Light
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="outline" color="dark" disabled>
-                Dark
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0"></CCol>
-          </CRow>
-        </CCardBody>
-      </CCard>
-      <CCard>
-        <CCardHeader>variant="ghost" Buttons</CCardHeader>
-        <CCardBody>
-          <p>
-            Use <code>.btn-ghost-*</code> class for variant="ghost" buttons.
-          </p>
-          <CRow className="align-items-center">
-            <CCol col="12" xl className="mb-3 mb-xl-0">
-              Normal
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="ghost" color="primary">
-                Primary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="ghost" color="secondary">
-                Secondary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="ghost" color="success">
-                Success
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="ghost" color="warning">
-                Warning
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="ghost" color="danger">
-                Danger
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="ghost" color="info">
-                Info
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="ghost" color="light">
-                Light
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="ghost" color="dark">
-                Dark
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0"></CCol>
-          </CRow>
-          <CRow className="align-items-center mt-3">
-            <CCol col="12" xl className="mb-3 mb-xl-0">
-              Active State
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                block
-                active
-                variant="ghost"
-                color="primary"
-                aria-pressed="true"
-              >
-                Primary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                block
-                active
-                variant="ghost"
-                color="secondary"
-                aria-pressed="true"
-              >
-                Secondary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                block
-                active
-                variant="ghost"
-                color="success"
-                aria-pressed="true"
-              >
-                Success
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                block
-                active
-                variant="ghost"
-                color="warning"
-                aria-pressed="true"
-              >
-                Warning
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                block
-                active
-                variant="ghost"
-                color="danger"
-                aria-pressed="true"
-              >
-                Danger
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                block
-                active
-                variant="ghost"
-                color="info"
-                aria-pressed="true"
-              >
-                Info
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                block
-                active
-                variant="ghost"
-                color="light"
-                aria-pressed="true"
-              >
-                Light
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                block
-                active
-                variant="ghost"
-                color="dark"
-                aria-pressed="true"
-              >
-                Dark
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0"></CCol>
-          </CRow>
-          <CRow className="align-items-center mt-3">
-            <CCol col="12" xl className="mb-3 mb-xl-0">
-              Disabled
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="ghost" color="primary" disabled>
-                Primary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="ghost" color="secondary" disabled>
-                Secondary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="ghost" color="success" disabled>
-                Success
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="ghost" color="warning" disabled>
-                Warning
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="ghost" color="danger" disabled>
-                Danger
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="ghost" color="info" disabled>
-                Info
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="ghost" color="light" disabled>
-                Light
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block variant="ghost" color="dark" disabled>
-                Dark
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0"></CCol>
-          </CRow>
-        </CCardBody>
-      </CCard>
-      <CCard>
-        <CCardHeader>Square Buttons</CCardHeader>
-        <CCardBody>
-          <p>
-            Use <code>.btn-square</code> class for square buttons.
-          </p>
-          <CRow className="align-items-center">
-            <CCol col="12" xl className="mb-3 mb-xl-0">
-              Normal
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="square" color="primary">
-                Primary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="square" color="secondary">
-                Secondary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="square" color="success">
-                Success
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="square" color="warning">
-                Warning
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="square" color="danger">
-                Danger
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="square" color="info">
-                Info
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="square" color="light">
-                Light
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="square" color="dark">
-                Dark
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="square" color="link">
-                Link
-              </CButton>
-            </CCol>
-          </CRow>
-          <CRow className="align-items-center mt-3">
-            <CCol col="12" xl className="mb-3 mb-xl-0">
-              Active State
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                active
-                block
-                shape="square"
-                color="primary"
-                aria-pressed="true"
-              >
-                Primary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                active
-                block
-                shape="square"
-                color="secondary"
-                aria-pressed="true"
-              >
-                Secondary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                active
-                block
-                shape="square"
-                color="success"
-                aria-pressed="true"
-              >
-                Success
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                active
-                block
-                shape="square"
-                color="warning"
-                aria-pressed="true"
-              >
-                Warning
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                active
-                block
-                shape="square"
-                color="danger"
-                aria-pressed="true"
-              >
-                Danger
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                active
-                block
-                shape="square"
-                color="info"
-                aria-pressed="true"
-              >
-                Info
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                active
-                block
-                shape="square"
-                color="light"
-                aria-pressed="true"
-              >
-                Light
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                active
-                block
-                shape="square"
-                color="dark"
-                aria-pressed="true"
-              >
-                Dark
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                active
-                block
-                shape="square"
-                color="link"
-                aria-pressed="true"
-              >
-                Link
-              </CButton>
-            </CCol>
-          </CRow>
-          <CRow className="align-items-center mt-3">
-            <CCol col="12" xl className="mb-3 mb-xl-0">
-              Disabled
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="square" color="primary" disabled>
-                Primary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="square" color="secondary" disabled>
-                Secondary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="square" color="success" disabled>
-                Success
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="square" color="warning" disabled>
-                Warning
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="square" color="danger" disabled>
-                Danger
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="square" color="info" disabled>
-                Info
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="square" color="light" disabled>
-                Light
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="square" color="dark" disabled>
-                Dark
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="square" color="link" disabled>
-                Link
-              </CButton>
-            </CCol>
-          </CRow>
-        </CCardBody>
-      </CCard>
-      <CCard>
-        <CCardHeader>Pill Buttons</CCardHeader>
-        <CCardBody>
-          <p>
-            Use <code>.btn-pill</code> class for pill buttons.
-          </p>
-          <CRow className="align-items-center">
-            <CCol col="12" xl className="mb-3 mb-xl-0">
-              Normal
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="pill" color="primary" className="">
-                Primary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="pill" color="secondary">
-                Secondary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="pill" color="success">
-                Success
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="pill" color="warning">
-                Warning
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="pill" color="danger">
-                Danger
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="pill" color="info">
-                Info
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="pill" color="light">
-                Light
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="pill" color="dark">
-                Dark
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="pill" color="link">
-                Link
-              </CButton>
-            </CCol>
-          </CRow>
-          <CRow className="align-items-center mt-3">
-            <CCol col="12" xl className="mb-3 mb-xl-0">
-              Active State
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                active
-                block
-                shape="pill"
-                color="primary"
-                aria-pressed="true"
-              >
-                Primary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                active
-                block
-                shape="pill"
-                color="secondary"
-                aria-pressed="true"
-              >
-                Secondary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                active
-                block
-                shape="pill"
-                color="success"
-                aria-pressed="true"
-              >
-                Success
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                active
-                block
-                shape="pill"
-                color="warning"
-                aria-pressed="true"
-              >
-                Warning
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                active
-                block
-                shape="pill"
-                color="danger"
-                aria-pressed="true"
-              >
-                Danger
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                active
-                block
-                shape="pill"
-                color="info"
-                aria-pressed="true"
-              >
-                Info
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                active
-                block
-                shape="pill"
-                color="light"
-                aria-pressed="true"
-              >
-                Light
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                active
-                block
-                shape="pill"
-                color="dark"
-                aria-pressed="true"
-              >
-                Dark
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton
-                active
-                block
-                shape="pill"
-                color="link"
-                aria-pressed="true"
-              >
-                Link
-              </CButton>
-            </CCol>
-          </CRow>
-          <CRow className="align-items-center mt-3">
-            <CCol col="12" xl className="mb-3 mb-xl-0">
-              Disabled
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="pill" color="primary" disabled>
-                Primary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="pill" color="secondary" disabled>
-                Secondary
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="pill" color="success" disabled>
-                Success
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="pill" color="warning" disabled>
-                Warning
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="pill" color="danger" disabled>
-                Danger
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="pill" color="info" disabled>
-                Info
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="pill" color="light" disabled>
-                Light
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="pill" color="dark" disabled>
-                Dark
-              </CButton>
-            </CCol>
-            <CCol col="6" sm="4" md="2" xl className="mb-3 mb-xl-0">
-              <CButton block shape="pill" color="link" disabled>
-                Link
-              </CButton>
-            </CCol>
-          </CRow>
-        </CCardBody>
-      </CCard>
-
-      <CCard>
-        <CCardHeader>Sizes</CCardHeader>
-        <CCardBody>
-          <p>
-            Fancy larger or smaller buttons? Add <code>size="lg"</code> or{' '}
-            <code>size="sm"</code> for additional sizes.
-          </p>
-          <CRow className="align-items-center">
-            <CCol col="2" xl className="mb-3 mb-xl-0">
-              Small
-            </CCol>
-            <CCol col="2" className="mb-3 mb-xl-0 text-center">
-              <CButton color="primary" size="sm">
-                Standard Button
-              </CButton>
-            </CCol>
-            <CCol col="2" className="mb-3 mb-xl-0 text-center">
-              <CButton color="secondary" variant="outline" size="sm">
-                Outline Button
-              </CButton>
-            </CCol>
-            <CCol col="2" className="mb-3 mb-xl-0 text-center">
-              <CButton size="sm" variant="ghost" color="ghost">
-                Ghost Button
-              </CButton>
-            </CCol>
-            <CCol col="2" className="mb-3 mb-xl-0 text-center">
-              <CButton shape="square" color="warning" size="sm">
-                Square Button
-              </CButton>
-            </CCol>
-            <CCol col="2" className="mb-3 mb-xl-0 text-center">
-              <CButton shape="pill" color="danger" size="sm">
-                Pill Button
-              </CButton>
-            </CCol>
-          </CRow>
-          <CRow className="align-items-center mt-3">
-            <CCol col="2" xl className="mb-3 mb-xl-0">
-              Normal
-            </CCol>
-            <CCol col="2" className="mb-3 mb-xl-0 text-center">
-              <CButton color="primary">Standard Button</CButton>
-            </CCol>
-            <CCol col="2" className="mb-3 mb-xl-0 text-center">
-              <CButton variant="outline" color="secondary">
-                Outline Button
-              </CButton>
-            </CCol>
-            <CCol col="2" className="mb-3 mb-xl-0 text-center">
-              <CButton variant="ghost" color="success">
-                Ghost Button
-              </CButton>
-            </CCol>
-            <CCol col="2" className="mb-3 mb-xl-0 text-center">
-              <CButton shape="square" color="warning">
-                Square Button
-              </CButton>
-            </CCol>
-            <CCol col="2" className="mb-3 mb-xl-0 text-center">
-              <CButton shape="pill" color="danger">
-                Pill Button
-              </CButton>
-            </CCol>
-          </CRow>
-          <CRow className="align-items-center mt-3">
-            <CCol col="2" xl className="mb-3 mb-xl-0">
-              Large
-            </CCol>
-            <CCol col="2" className="mb-3 mb-xl-0 text-center">
+              <CButton type="submit" color="primary">
+                Button
+              </CButton>
+              <CButton component="input" type="button" color="primary" value="Input" />
+              <CButton component="input" type="submit" color="primary" value="Submit" />
+              <CButton component="input" type="reset" color="primary" value="Reset" />
+            </DocsExample>
+          </CCardBody>
+        </CCard>
+      </CCol>
+      <CCol xs={12}>
+        <CCard className="mb-4">
+          <CCardHeader>
+            <strong>React Button</strong> <small>outline</small>
+          </CCardHeader>
+          <CCardBody>
+            <p className="text-medium-emphasis small">
+              If you need a button, but without the strong background colors. Set{' '}
+              <code>variant=&#34;outline&#34;</code> prop to remove all background colors.
+            </p>
+            <DocsExample href="components/buttons#outline-buttons">
+              {['normal', 'active', 'disabled'].map((state, index) => (
+                <CRow className="align-items-center mb-3" key={index}>
+                  <CCol xs={12} xl={2} className="mb-3 mb-xl-0">
+                    {state.charAt(0).toUpperCase() + state.slice(1)}
+                  </CCol>
+                  <CCol xs>
+                    {[
+                      'primary',
+                      'secondary',
+                      'success',
+                      'danger',
+                      'warning',
+                      'info',
+                      'light',
+                      'dark',
+                    ].map((color, index) => (
+                      <CButton
+                        color={color}
+                        variant="outline"
+                        key={index}
+                        active={state === 'active'}
+                        disabled={state === 'disabled'}
+                      >
+                        {color.charAt(0).toUpperCase() + color.slice(1)}
+                      </CButton>
+                    ))}
+                  </CCol>
+                </CRow>
+              ))}
+            </DocsExample>
+          </CCardBody>
+        </CCard>
+      </CCol>
+      <CCol xs={12}>
+        <CCard className="mb-4">
+          <CCardHeader>
+            <strong>React Button</strong> <small>ghost</small>
+          </CCardHeader>
+          <CCardBody>
+            <p className="text-medium-emphasis small">
+              If you need a ghost variant of button, set <code>variant=&#34;ghost&#34;</code> prop
+              to remove all background colors.
+            </p>
+            <DocsExample href="components/buttons#ghost-buttons">
+              {['normal', 'active', 'disabled'].map((state, index) => (
+                <CRow className="align-items-center mb-3" key={index}>
+                  <CCol xs={12} xl={2} className="mb-3 mb-xl-0">
+                    {state.charAt(0).toUpperCase() + state.slice(1)}
+                  </CCol>
+                  <CCol xs>
+                    {[
+                      'primary',
+                      'secondary',
+                      'success',
+                      'danger',
+                      'warning',
+                      'info',
+                      'light',
+                      'dark',
+                    ].map((color, index) => (
+                      <CButton
+                        color={color}
+                        variant="ghost"
+                        key={index}
+                        active={state === 'active'}
+                        disabled={state === 'disabled'}
+                      >
+                        {color.charAt(0).toUpperCase() + color.slice(1)}
+                      </CButton>
+                    ))}
+                  </CCol>
+                </CRow>
+              ))}
+            </DocsExample>
+          </CCardBody>
+        </CCard>
+      </CCol>
+      <CCol xs={12}>
+        <CCard className="mb-4">
+          <CCardHeader>
+            <strong>React Button</strong> <small>Sizes</small>
+          </CCardHeader>
+          <CCardBody>
+            <p className="text-medium-emphasis small">
+              Larger or smaller buttons? Add <code>size=&#34;lg&#34;</code> or{' '}
+              <code>size=&#34;sm&#34;</code> for additional sizes.
+            </p>
+            <DocsExample href="components/buttons#sizes">
               <CButton color="primary" size="lg">
-                Standard Button
+                Large button
               </CButton>
-            </CCol>
-            <CCol col="2" className="mb-3 mb-xl-0 text-center">
-              <CButton variant="outline" color="secondary" size="lg">
-                Outline Button
+              <CButton color="secondary" size="lg">
+                Large button
               </CButton>
-            </CCol>
-            <CCol col="2" className="mb-3 mb-xl-0 text-center">
-              <CButton variant="ghost" color="success" size="lg">
-                Ghost Button
+            </DocsExample>
+            <DocsExample href="components/buttons#sizes">
+              <CButton color="primary" size="sm">
+                Small button
               </CButton>
-            </CCol>
-            <CCol col="2" className="mb-3 mb-xl-0 text-center">
-              <CButton shape="square" color="warning" size="lg">
-                Square Button
+              <CButton color="secondary" size="sm">
+                Small button
               </CButton>
-            </CCol>
-            <CCol col="2" className="mb-3 mb-xl-0 text-center">
-              <CButton shape="pill" color="danger" size="lg">
-                Pill Button
+            </DocsExample>
+          </CCardBody>
+        </CCard>
+      </CCol>
+      <CCol xs={12}>
+        <CCard className="mb-4">
+          <CCardHeader>
+            <strong>React Button</strong> <small>Pill</small>
+          </CCardHeader>
+          <CCardBody>
+            <DocsExample href="components/buttons#pill-buttons">
+              {[
+                'primary',
+                'secondary',
+                'success',
+                'danger',
+                'warning',
+                'info',
+                'light',
+                'dark',
+              ].map((color, index) => (
+                <CButton color={color} shape="rounded-pill" key={index}>
+                  {color.charAt(0).toUpperCase() + color.slice(1)}
+                </CButton>
+              ))}
+            </DocsExample>
+          </CCardBody>
+        </CCard>
+      </CCol>
+      <CCol xs={12}>
+        <CCard className="mb-4">
+          <CCardHeader>
+            <strong>React Button</strong> <small>Square</small>
+          </CCardHeader>
+          <CCardBody>
+            <DocsExample href="components/buttons#square">
+              {[
+                'primary',
+                'secondary',
+                'success',
+                'danger',
+                'warning',
+                'info',
+                'light',
+                'dark',
+              ].map((color, index) => (
+                <CButton color={color} shape="rounded-0" key={index}>
+                  {color.charAt(0).toUpperCase() + color.slice(1)}
+                </CButton>
+              ))}
+            </DocsExample>
+          </CCardBody>
+        </CCard>
+      </CCol>
+      <CCol xs={12}>
+        <CCard className="mb-4">
+          <CCardHeader>
+            <strong>React Button</strong> <small>Disabled state</small>
+          </CCardHeader>
+          <CCardBody>
+            <p className="text-medium-emphasis small">
+              Add the <code>disabled</code> boolean prop to any <code>&lt;CButton&gt;</code>{' '}
+              component to make buttons look inactive. Disabled button has{' '}
+              <code>pointer-events: none</code> applied to, disabling hover and active states from
+              triggering.
+            </p>
+            <DocsExample href="components/buttons#disabled-state">
+              <CButton color="primary" size="lg" disabled>
+                Primary button
               </CButton>
-            </CCol>
-          </CRow>
-        </CCardBody>
-      </CCard>
-
-      <CCard>
-        <CCardHeader>With Icons</CCardHeader>
-        <CCardBody>
-          <CRow className="align-items-center mt-3">
-            <CCol col="2" className="text-center mt-3">
-              <CButton color="primary">Standard Button</CButton>
-            </CCol>
-            <CCol col="2" className="text-center mt-3">
-              <CButton color="secondary" variant="outline">
-                <CIcon name="cil-lightbulb" />
-                Outline Button
+              <CButton color="secondary" size="lg" disabled>
+                Button
               </CButton>
-            </CCol>
-            <CCol col="2" className="text-center mt-3">
-              <CButton variant="ghost" color="success">
-                <CIcon name="cil-lightbulb" />
-                Ghost Button
+            </DocsExample>
+            <p className="text-medium-emphasis small">
+              Disabled buttons using the <code>&lt;a&gt;</code> component act a little different:
+            </p>
+            <p className="text-medium-emphasis small">
+              <code>&lt;a&gt;</code>s don&#39;tsupport the <code>disabled</code> attribute, so
+              CoreUI has to add <code>.disabled</code> className to make buttons look inactive.
+              CoreUI also has to add to the disabled button component{' '}
+              <code>aria-disabled=&#34;true&#34;</code> attribute to show the state of the component
+              to assistive technologies.
+            </p>
+            <DocsExample href="components/buttons#disabled-state">
+              <CButton component="a" href="#" color="primary" size="lg" disabled>
+                Primary link
               </CButton>
-            </CCol>
-            <CCol col="2" className="text-center mt-3">
-              <CButton shape="square" color="warning">
-                <CIcon name="cil-lightbulb" />
-                Square Button
+              <CButton component="a" href="#" color="secondary" size="lg" disabled>
+                Link
               </CButton>
-            </CCol>
-            <CCol col="2" className="text-center mt-3">
-              <CButton shape="pill" color="danger">
-                <CIcon name="cil-lightbulb" />
-                Pill Button
-              </CButton>
-            </CCol>
-          </CRow>
-        </CCardBody>
-      </CCard>
-      <CRow>
-        <CCol xs="12" md="6">
-          <CCard>
-            <CCardHeader>Block Level Buttons</CCardHeader>
-            <CCardBody>
-              <p>
-                Add prop <code>block</code>
-              </p>
-              <CButton color="secondary" size="lg" block>
-                Block level button
-              </CButton>
-              <CButton color="primary" size="lg" block>
-                Block level button
-              </CButton>
-              <CButton color="success" size="lg" block>
-                Block level button
-              </CButton>
-              <CButton color="info" size="lg" block>
-                Block level button
-              </CButton>
-              <CButton color="warning" size="lg" block>
-                Block level button
-              </CButton>
-              <CButton color="danger" size="lg" block>
-                Block level button
-              </CButton>
-              <CButton color="link" size="lg" block>
-                Block level button
-              </CButton>
-            </CCardBody>
-          </CCard>
-        </CCol>
-        <CCol xs="12" md="6">
-          <CCard>
-            <CCardHeader>Block Level Buttons</CCardHeader>
-            <CCardBody>
-              <p>
-                Add prop <code>block</code>
-              </p>
-              <CButton variant="outline" color="secondary" size="lg" block>
-                Block level button
-              </CButton>
-              <CButton variant="outline" color="primary" size="lg" block>
-                Block level button
-              </CButton>
-              <CButton variant="outline" color="success" size="lg" block>
-                Block level button
-              </CButton>
-              <CButton variant="outline" color="info" size="lg" block>
-                Block level button
-              </CButton>
-              <CButton variant="outline" color="warning" size="lg" block>
-                Block level button
-              </CButton>
-              <CButton variant="outline" color="danger" size="lg" block>
-                Block level button
-              </CButton>
-              <CButton variant="ghost" color="info" size="lg" block>
-                Block level button
-              </CButton>
-            </CCardBody>
-          </CCard>
-        </CCol>
-      </CRow>
-    </>
+            </DocsExample>
+          </CCardBody>
+        </CCard>
+      </CCol>
+      <CCol xs={12}>
+        <CCard className="mb-4">
+          <CCardHeader>
+            <strong>React Button</strong> <small>Block buttons</small>
+          </CCardHeader>
+          <CCardBody>
+            <p className="text-medium-emphasis small">
+              Create buttons that span the full width of a parent—by using utilities.
+            </p>
+            <DocsExample href="components/buttons#block-buttons">
+              <div className="d-grid gap-2">
+                <CButton color="primary">Button</CButton>
+                <CButton color="primary">Button</CButton>
+              </div>
+            </DocsExample>
+            <p className="text-medium-emphasis small">
+              Here we create a responsive variation, starting with vertically stacked buttons until
+              the <code>md</code> breakpoint, where <code>.d-md-block</code> replaces the{' '}
+              <code>.d-grid</code> class, thus nullifying the <code>gap-2</code> utility. Resize
+              your browser to see them change.
+            </p>
+            <DocsExample href="components/buttons#block-buttons">
+              <div className="d-grid gap-2 d-md-block">
+                <CButton color="primary">Button</CButton>
+                <CButton color="primary">Button</CButton>
+              </div>
+            </DocsExample>
+            <p className="text-medium-emphasis small">
+              You can adjust the width of your block buttons with grid column width classes. For
+              example, for a half-width &#34;block button&#34;, use <code>.col-6</code>. Center it
+              horizontally with <code>.mx-auto</code>, too.
+            </p>
+            <DocsExample href="components/buttons#block-buttons">
+              <div className="d-grid gap-2 col-6 mx-auto">
+                <CButton color="primary">Button</CButton>
+                <CButton color="primary">Button</CButton>
+              </div>
+            </DocsExample>
+            <p className="text-medium-emphasis small">
+              Additional utilities can be used to adjust the alignment of buttons when horizontal.
+              Here we&#39;ve taken our previous responsive example and added some flex utilities and
+              a margin utility on the button to right align the buttons when they&#39;re no longer
+              stacked.
+            </p>
+            <DocsExample href="components/buttons#block-buttons">
+              <div className="d-grid gap-2 d-md-flex justify-content-md-end">
+                <CButton color="primary" className="me-md-2">
+                  Button
+                </CButton>
+                <CButton color="primary">Button</CButton>
+              </div>
+            </DocsExample>
+          </CCardBody>
+        </CCard>
+      </CCol>
+    </CRow>
   )
 }
 
