@@ -3,42 +3,33 @@ import {
   CButton,
   CCol,
   CContainer,
-  CInput,
+  CFormInput,
   CInputGroup,
-  CInputGroupAppend,
-  CInputGroupPrepend,
   CInputGroupText,
   CRow,
 } from '@coreui/react'
 import CIcon from '@coreui/icons-react'
+import { cilMagnifyingGlass } from '@coreui/icons'
 
 const Page500 = () => {
   return (
-    <div className="c-app c-default-layout flex-row align-items-center">
+    <div className="bg-light min-vh-100 d-flex flex-row align-items-center">
       <CContainer>
         <CRow className="justify-content-center">
-          <CCol md="6">
+          <CCol md={6}>
             <span className="clearfix">
-              <h1 className="float-left display-3 mr-4">500</h1>
+              <h1 className="float-start display-3 me-4">500</h1>
               <h4 className="pt-3">Houston, we have a problem!</h4>
-              <p className="text-muted float-left">
+              <p className="text-medium-emphasis float-start">
                 The page you are looking for is temporarily unavailable.
               </p>
             </span>
             <CInputGroup className="input-prepend">
-              <CInputGroupPrepend>
-                <CInputGroupText>
-                  <CIcon name="cil-magnifying-glass" />
-                </CInputGroupText>
-              </CInputGroupPrepend>
-              <CInput
-                size="16"
-                type="text"
-                placeholder="What are you looking for?"
-              />
-              <CInputGroupAppend>
-                <CButton color="info">Search</CButton>
-              </CInputGroupAppend>
+              <CInputGroupText>
+                <CIcon icon={cilMagnifyingGlass} />
+              </CInputGroupText>
+              <CFormInput type="text" placeholder="What are you looking for?" />
+              <CButton color="info">Search</CButton>
             </CInputGroup>
           </CCol>
         </CRow>
