@@ -1,7 +1,9 @@
+import PropTypes from 'prop-types'
 import React, { useEffect, useState, createRef } from 'react'
 import classNames from 'classnames'
-import { CRow, CCol } from '@coreui/react'
-import { rgbToHex } from '@coreui/utils/src'
+import { CRow, CCol, CCard, CCardHeader, CCardBody } from '@coreui/react'
+import { rgbToHex } from '@coreui/utils'
+import { DocsLink } from 'src/components'
 
 const ThemeView = () => {
   const [color, setColor] = useState('rgb(255, 255, 255)')
@@ -9,9 +11,7 @@ const ThemeView = () => {
 
   useEffect(() => {
     const el = ref.current.parentNode.firstChild
-    const varColor = window
-      .getComputedStyle(el)
-      .getPropertyValue('background-color')
+    const varColor = window.getComputedStyle(el).getPropertyValue('background-color')
     setColor(varColor)
   }, [ref])
 
@@ -19,11 +19,11 @@ const ThemeView = () => {
     <table className="table w-100" ref={ref}>
       <tbody>
         <tr>
-          <td className="text-muted">HEX:</td>
+          <td className="text-medium-emphasis">HEX:</td>
           <td className="font-weight-bold">{rgbToHex(color)}</td>
         </tr>
         <tr>
-          <td className="text-muted">RGB:</td>
+          <td className="text-medium-emphasis">RGB:</td>
           <td className="font-weight-bold">{color}</td>
         </tr>
       </tbody>
@@ -34,7 +34,7 @@ const ThemeView = () => {
 const ThemeColor = ({ className, children }) => {
   const classes = classNames(className, 'theme-color w-75 rounded mb-3')
   return (
-    <CCol xl="2" md="4" sm="6" xs="12" className="mb-4">
+    <CCol xs={12} sm={6} md={4} xl={2} className="mb-4">
       <div className={classes} style={{ paddingTop: '75%' }}></div>
       {children}
       <ThemeView />
@@ -42,12 +42,20 @@ const ThemeColor = ({ className, children }) => {
   )
 }
 
+ThemeColor.propTypes = {
+  children: PropTypes.node,
+  className: PropTypes.string,
+}
+
 const Colors = () => {
   return (
     <>
-      <div className="card">
-        <div className="card-header">Theme colors</div>
-        <div className="card-body">
+      <CCard className="mb-4">
+        <CCardHeader>
+          Theme colors
+          <DocsLink href="https://coreui.io/docs/utilities/colors/" />
+        </CCardHeader>
+        <CCardBody>
           <CRow>
             <ThemeColor className="bg-primary">
               <h6>Brand Primary Color</h6>
@@ -74,42 +82,8 @@ const Colors = () => {
               <h6>Brand Dark Color</h6>
             </ThemeColor>
           </CRow>
-        </div>
-      </div>
-      <div className="card">
-        <div className="card-header">Grays</div>
-        <div className="card-body">
-          <CRow className="mb-3">
-            <ThemeColor className="bg-gray-100">
-              <h6>Gray 100 Color</h6>
-            </ThemeColor>
-            <ThemeColor className="bg-gray-200">
-              <h6>Gray 200 Color</h6>
-            </ThemeColor>
-            <ThemeColor className="bg-gray-300">
-              <h6>Gray 300 Color</h6>
-            </ThemeColor>
-            <ThemeColor className="bg-gray-400">
-              <h6>Gray 400 Color</h6>
-            </ThemeColor>
-            <ThemeColor className="bg-gray-500">
-              <h6>Gray 500 Color</h6>
-            </ThemeColor>
-            <ThemeColor className="bg-gray-600">
-              <h6>Gray 600 Color</h6>
-            </ThemeColor>
-            <ThemeColor className="bg-gray-700">
-              <h6>Gray 700 Color</h6>
-            </ThemeColor>
-            <ThemeColor className="bg-gray-800">
-              <h6>Gray 800 Color</h6>
-            </ThemeColor>
-            <ThemeColor className="bg-gray-900">
-              <h6>Gray 900 Color</h6>
-            </ThemeColor>
-          </CRow>
-        </div>
-      </div>
+        </CCardBody>
+      </CCard>
     </>
   )
 }
